a second from CLI is stretched from 1000ms to 1024ms for easier calculations inside the PIC. In a next step the interface between Client and PIC should be modified to read the timevalue N as N= 1024ms * t
reduced complexity in set fade. We allow only fading in steps of 16, this is nearly not visible to the human eye, but reduces calculation.
why 16? because stepSize * CYCLE_TMMS is the minimal garanteed fading interval = 1024ms
now, we have only one division: periodeLength = FadeTmms/numberSteps
periodeLength is the number of CYCLES have to left until one fading step is executed.