linkit: add ttyS as SPI sink
On LinkIt 7688 Duo WyLight firmware runs on the OpenWRT mips MPU, which
is connected to an ATmega32U4 MCU. MPU<->MCU are connected by a serial
line.
SPI and PWM of the MCU are used for output. spi_ttyS.c is the driver
running on MPU sending the LEDStrip states as byte array over serial to
the MCU.