cc3200: begin complete removal of cc3200
cc3200 became to expensive and has no advantages over linkit7688.
As the CC3200 port is to complex, we decided to remove cc3200 support
completely. However the branch obsolete-cc3200 should remain, to keep
easy access to the cc3200 implementation, if it might become of
interest again.

This commit starts the removal by simple deleting cc3200 only files.